# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a Manga Scraper API server.",
        "title": "Manga Scraper API",
        "termsOfService": "https://manga-scraper.hostinger.fourleaves.studio/terms",
        "contact": {
            "name": "API Support",
            "url": "https://manga-scraper.hostinger.fourleaves.studio/support",
            "email": "admin@fourleaves.studio"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/chapters/{provider_slug}/{series_slug}/{chapter_slug}": {
            "get": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Get chapter by slug",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "chapters"
                ],
                "summary": "Get chapter by slug",
                "parameters": [
                    {
                        "type": "string",
                        "example": "asura",
                        "description": "Provider slug",
                        "name": "provider_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "reincarnator",
                        "description": "Series slug",
                        "name": "series_slug",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "reincarnator-chapter-0",
                        "description": "Chapter slug",
                        "name": "chapter_slug",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scrape-requests/chapters/list": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create request to scrape chapter list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape-requests"
                ],
                "summary": "Create request to scrape chapter list",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Binding.PostScrapeChapterList"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
        "/api/v1/scrape-requests/series/list": {
            "post": {
                "security": [
                    {
                        "TokenAuth": []
                    }
                ],
                "description": "Create request to scrape series list",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "scrape-requests"
                ],
                "summary": "Create request to scrape series list",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/v1Binding.PostScrapeSeriesList"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/v1Response.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "v1Binding.PostScrapeChapterList": {
            "type": "object",
            "required": [
                "provider",
                "series"
            ],
            "properties": {
                "provider": {
                    "type": "string",
                    "example": "asura"
                },
                "series": {
                    "type": "string",
                    "example": "reincarnator"
                }
            }
        },
        "v1Binding.PostScrapeSeriesList": {
            "type": "object",
            "required": [
                "provider"
            ],
            "properties": {
                "provider": {
                    "type": "string",
                    "example": "asura"
                }
            }
        },
...
        "v1Response.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "detail": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "TokenAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}